### 2022-04-27 #4352 @editorialbot
[PRE REVIEW]: CiteLang: Modeling the Research Software Ecosystem
**Submitting author:** @vsoch (<a href="http://orcid.org/0000-0002-4387-3819">Vanessa Sochat</a>)
**Repository:** https://github.com/vsoch/citelang
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.27
