### 2022-04-29 8857a90
Remove doi prefix from paper
