+++ paper/paper.bib
@@ -40,7 +40,7 @@ number = {8},
 pages = {491-506},
 year = {2002},
 issn = {0950-5849},
-doi = {https://doi.org/10.1016/S0950-5849(02)00049-6},
+doi = {10.1016/S0950-5849(02)00049-6},
 url = {https://www.sciencedirect.com/science/article/pii/S0950584902000496},
 author = {R.L. Glass and I. Vessey and V. Ramesh},
 keywords = {Topic: computing research, Research approach: evaluative-other, Research method: literature analysis, Reference discipline: not applicable, Level of analysis: profession},
