### 2022-05-15 3e310cd
early work to add CMakeLists.txt parser!
we can actually get a ton of dependencies for Cpp via spack,
yay, even if C++ does not technically have its own package manager

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
