### 2022-05-20 #4352 @faroit
@Kevin-Mattheus-Moerman I spent some time getting into this now and - yes, while its clearly outside my core domain – I find the contribution interesting and relevant. Happy to accept
