### 2022-05-20 #4352 @faroit
👋  Hi @vsoch, thanks again for your submission to JOSS and sorry for the delay. I will be serving as the editor for this now. We'll use this issue to identify reviewers and resolve technical issues.

This tool implements a broad (and alternative) vision of how research software can be cited outside classical DOIs. I would like to diversify the reviewers in order to accommodate this vision but also assess the implementation details at the same time. E.g. at least one very experienced senior researcher that can comment on the broader vision of citelang. But also one experienced python expert.

I think the paper is already in a good shape. Looking at the repository however, one thing I clearly miss is more polished examples of typical outputs from popular repositories in order to better understand the vision of citelang and how it is recommended to be used by research software engineers.

I feel that potential reviewers would like to quickly assess what citelang is addressing for their citation/research software workflow. Therefore I would suggest - before we find potential reviewers to update the repository to accommodate:

- more examples outputs, ideally in a separate docs gallery.
- a more detailed "getting started" guide that describes the vision and how citelang addresses the issues raised in the JOSS paper. Ideally an example workflow.
- clarify why the software is subtitled "markdown syntax for citing software" - what is the relation to markdown?

@vsoch let me know if you think this sounds appropriate to you
