### 2022-05-20 f9d8120
adding --deep option to skip --first-parent
a git repo with a more complex history (e.g., squash and merge may miss
contributions by way of only looking at the first parent. This adds --deep
to remove that flag and show the full contribution line count

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
