+++ citelang/main/contrib.py
@@ -29,7 +29,7 @@ def blame_task(args):
     with blame, record the file as empty in the cache so we don't try again.
     """
     # Sploot out args into variables
-    root, commit, path, save_to = args
+    root, commit, path, save_to, dups = args
 
     # If we have the history already, return
     if os.path.exists(save_to):
@@ -43,21 +43,16 @@ def blame_task(args):
     # Atomic save
     tmp_file = save_to + ".tmp.%d" % os.getpid()
 
+    # Asssemble command
+    command = ["git", "blame", "-w", "--line-porcelain"]
+
+    # find line copies and movements across files (more shallow result)
+    if dups:
+        command += ["-M", "-C"]
+    command += [commit, "--", path]
     with utils.workdir(root):
         try:
-            res = utils.run_command(
-                [
-                    "git",
-                    "blame",
-                    "-w",
-                    "-M",
-                    "-C",
-                    "--line-porcelain",
-                    commit,
-                    "--",
-                    path,
-                ]
-            )
+            res = utils.run_command(command)
         # This means we hit some binary, etc.
         except:
             return {"empty": "%s-%s" % (commit, path)}
@@ -139,11 +134,12 @@ class CommitStats(GitParser):
     Stats for a particular commit, saved to a cache.
     """
 
-    def __init__(self, root, commit, paths, outdir):
+    def __init__(self, root, commit, paths, outdir, dups=True):
         self.paths = paths
         self.commit = commit
         self.items = {}
         self.root = root
+        self.dups = dups
 
         # This is a shared cache of blame output
         self.outdir = outdir
@@ -185,6 +181,7 @@ class CommitStats(GitParser):
                 self.commit,
                 f,
                 os.path.join(self.outdir, f, f"{self.commit}.json"),
+                self.dups,
             )
             for f in set(files)
             if "%s-%s" % (self.commit, f) not in self.empties
@@ -403,6 +400,10 @@ class ContributionParser(GitParser):
         prefixes = prefixes or []
         regex = "(%s)" % "|".join(prefixes) if prefixes else ""
 
+        if not prefixes:
+            self.paths = []
+            return
+
         # Get all files in the commit history
         for path in self.git(
             "git", "log", "--pretty=format:", "--name-only", "--diff-filter=A"
@@ -430,7 +431,7 @@ class ContributionParser(GitParser):
         commits = self.get_commit_range(first_parent)
 
         # Retrieve items of history
-        history = self.index_history(commits)
+        history = self.index_history(commits, dups=first_parent)
 
         # Do we want to only include within a range?
         start_timestamp = (
@@ -451,9 +452,9 @@ class ContributionParser(GitParser):
         """
         return int(self.git("git", "show", "-s", "--format=%ct", commit))
 
-    def index_history(self, commits):
+    def index_history(self, commits, dups=True):
         """
-        index git history. We should eventually run this in parallel.
+        index git history.
         """
         # Keep a lookup of commit stats by directory
         parsers = {}
@@ -480,7 +481,7 @@ class ContributionParser(GitParser):
             total = 0
             print("Preparing all tasks...", end="\r")
             for commit in commits:
-                parser = CommitStats(self.root, commit, self.paths, cache)
+                parser = CommitStats(self.root, commit, self.paths, cache, dups=dups)
                 print("Preparing all tasks...%s" % next(spins), end="\r")
                 total += parser.ntasks
                 parsers[commit] = parser
