### 2022-05-20 0d9fd50
default for contrib is now deep search
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
