+++ citelang/main/contrib.py
@@ -29,7 +29,7 @@ def blame_task(args):
     with blame, record the file as empty in the cache so we don't try again.
     """
     # Sploot out args into variables
-    root, commit, path, save_to, dups = args
+    root, commit, path, save_to, shallow = args
 
     # If we have the history already, return
     if os.path.exists(save_to):
@@ -47,7 +47,7 @@ def blame_task(args):
     command = ["git", "blame", "-w", "--line-porcelain"]
 
     # find line copies and movements across files (more shallow result)
-    if dups:
+    if shallow:
         command += ["-M", "-C"]
     command += [commit, "--", path]
     with utils.workdir(root):
@@ -134,12 +134,12 @@ class CommitStats(GitParser):
     Stats for a particular commit, saved to a cache.
     """
 
-    def __init__(self, root, commit, paths, outdir, dups=True):
+    def __init__(self, root, commit, paths, outdir, shallow=False):
         self.paths = paths
         self.commit = commit
         self.items = {}
         self.root = root
-        self.dups = dups
+        self.shallow = shallow
 
         # This is a shared cache of blame output
         self.outdir = outdir
@@ -181,7 +181,7 @@ class CommitStats(GitParser):
                 self.commit,
                 f,
                 os.path.join(self.outdir, f, f"{self.commit}.json"),
-                self.dups,
+                self.shallow,
             )
             for f in set(files)
             if "%s-%s" % (self.commit, f) not in self.empties
@@ -418,7 +418,7 @@ class ContributionParser(GitParser):
 
         self.paths = list(paths)
 
-    def parse(self, return_summary=True, within_range=True, first_parent=True):
+    def parse(self, return_summary=True, within_range=True, shallow=False):
         """
         Parse the contributions. If within_range is True, don't include git
         blame that goes outside of the range provided.
@@ -428,10 +428,10 @@ class ContributionParser(GitParser):
             logger.exit(f"Cannot find .git repo in {self.root}")
 
         # get commits associated with start to end
-        commits = self.get_commit_range(first_parent)
+        commits = self.get_commit_range(shallow)
 
         # Retrieve items of history
-        history = self.index_history(commits, dups=first_parent)
+        history = self.index_history(commits, shallow=shallow)
 
         # Do we want to only include within a range?
         start_timestamp = (
@@ -452,7 +452,7 @@ class ContributionParser(GitParser):
         """
         return int(self.git("git", "show", "-s", "--format=%ct", commit))
 
-    def index_history(self, commits, dups=True):
+    def index_history(self, commits, shallow=False):
         """
         index git history.
         """
@@ -481,7 +481,9 @@ class ContributionParser(GitParser):
             total = 0
             print("Preparing all tasks...", end="\r")
             for commit in commits:
-                parser = CommitStats(self.root, commit, self.paths, cache, dups=dups)
+                parser = CommitStats(
+                    self.root, commit, self.paths, cache, shallow=shallow
+                )
                 print("Preparing all tasks...%s" % next(spins), end="\r")
                 total += parser.ntasks
                 parsers[commit] = parser
@@ -512,11 +514,11 @@ class ContributionParser(GitParser):
         """
         return self.git("git", "rev-list", "-n", "1", tag)
 
-    def get_commit_range(self, first_parent=True):
+    def get_commit_range(self, shallow=False):
         """
         Given a start and end, parse and return the commits from git
         """
-        if first_parent:
+        if shallow:
             res = self.git("git", "log", "--first-parent", "--all", "--format=%H")
 
         # Possibly duplications but won't miss any commits
