### 2022-05-20 5a69722
Default is deep search with option for --shallow (#34)
* adding --shallow option to use --first-parent
