+++ citelang/tests/test_packages.py
@@ -35,9 +35,7 @@ here = os.path.dirname(os.path.abspath(__file__))
                 "eigen3",
                 "openmp",
                 "lapack",
-                "gnuconfig",
                 "python",
-                "cpio",
                 "curl",
                 "xpat",
                 "zlib",
