### 2022-05-20 #4352 @vsoch
hey @faroit totally no worries about the timing - it's an amazing effort that y'all do here to do this in your free time (and it's still pretty fast imho).

> I think the paper is already in a good shape. Looking at the repository however, one thing I clearly miss is more polished examples of typical outputs from popular repositories in order to better understand the vision of citelang and how it is recommended to be used by research software engineers.

I'm happy to add this, but did you not see the docs? https://vsoch.github.io/citelang/getting_started/user-guide.html. CiteLang is a fairly diverse tool and there is no "here is the one use case" or "this is exactly how you are supposed to us it" for it. It's main purpose is that it's an API for getting data about contributions, turning that into a graph structure, and the intention (and why it is research software) is that you can use these graphs to understand an ecosystem as I've done [with the rsepedia-analysis](https://rseng.github.io/rsepedia-analysis/). That entire thing is derived from CiteLang data, and akin to how a researcher would use numpy in their project to do something higher level, I used citelang. The render/graph/badge provided with the software are extra bells and whistles, or how I chose to use these graphs (and provide fun tools for the interested user). So let's chat about what you think is missing - granted that you saw that documentation that has links to:

- badges that show the entire credit tree (including deps). That is shown in the README and in the [docs here](https://vsoch.github.io/citelang/getting_started/user-guide.html#badge) for multiple variations.
- summarizing contributions, described [here](https://vsoch.github.io/citelang/getting_started/user-guide.html#contrib)
- interactive badges and credit trees [linked from docs](https://vsoch.github.io/citelang/_static/example/badge/treemap/index.html)
- The functionality that originally derived the name - a "markdown syntax for citations" there is a [pre](https://github.com/vsoch/citelang/blob/main/examples/pre-render.md) and [post](https://github.com/vsoch/citelang/blob/main/examples/post-render.md) rendering, where citelang reads in the pre, parses the citation and generates those same credit trees, and makes a nice table. That is also done automated and present in [the citelang repository](https://github.com/vsoch/citelang/blob/main/software-credit.md).

And the content of the paper is somewhat represented [here](https://vsoch.github.io/citelang/) and at the top of [here](https://vsoch.github.io/citelang/getting_started/index.html).

> I feel that potential reviewers would like to quickly assess what citelang is addressing for their citation/research software workflow. Therefore I would suggest - before we find potential reviewers - to update the repository to accommodate:

I don't think I'm saying this is yet a part of any research software citation workflow? I think right now our workflow is "make bibtex or a DOI with Zenodo and copy paste into a bibliography" and CiteLang is saying "this could be different." CiteLang is intended to be provided to empower people to model the space and think more about these ideas. To go back to the numpy example, numpy itself isn't explicitly stated in some workflow for a domain of analysis, but it's an underlying library that forms components that turn out to be. Let me know if that makes sense.

> more examples outputs, ideally in a separate html docs gallery.

So I'm happy to add what you think is missing, but I don't see that any of the functionality that is provided is missing. Could you tell me what you think is missing? Do you want some sort of a "quick summary" bullet list of things that link to the same examples?  Specifically what output are you looking for? I thought the requirement for JoSS was robust documentation, which I think I have with every example that the software can do, so I apologize if I missed an extra requirement (and I'd like to ask for clarity or a link to what you are looking for).

> more detailed "getting started" guide that describes the vision and how citelang addresses the issues raised in the JOSS paper. Ideally an example workflow. I would also think that part of [your blog post](https://vsoch.github.io/2022/citelang-contrib/) can be used as material.

Would this also be addressed by this summary listing? I think I present the tool as one idea for how we can influence the future of credit for software, but I don't make claims that there is some single command here that you run and "that's the answer." CiteLang has a lot of cool functionality and commands, but those are more examples of what I've built using these APIs and graphs. I'm hoping other researchers interested in software credit might do similar and make other cool ideas.

> clarification of why the software is subtitled "markdown syntax for citing software" - what is the relation to markdown?

This is the syntax shown in the links above that is shown with [render](https://vsoch.github.io/citelang/getting_started/user-guide.html#render).

And there are GitHub actions for [all of those things](https://vsoch.github.io/citelang/getting_started/user-guide.html#github-action) so if someone just wants to "generate a cool badge or a software credit graph from a paper they have written in markdown" that's probably the most direct usage, along with summarizing contributions. E.g., the singularity project is using CIteLang to be able to say thank you to contributions for each release, here is an example I was just checking with one of their developers:

![image](https://user-images.githubusercontent.com/814322/169588300-d89be6f9-ebe3-40d3-8b4d-ef55193389fb.png)

Thank you for your comments!
