### 2022-05-20 #4352 @faroit
@vsoch thanks for your detailed explanations.

> I'm happy to add this, but did you not see the docs? https://vsoch.github.io/citelang/getting_started/user-guide.html. CiteLang is a fairly diverse tool and there is no "here is the one use case" or "this is exactly how you are supposed to us it" for it.

Thanks for the clarification. In the submitted paper you state:

> "On a cultural level, CiteLang is the first tool of its kind to suggest an alternative way to value software beyond the traditional academic practice of publication. It suggests a model where RSEng are not required to generate separate DOIs (digital object identifiers) associated with papers..."

That id why, when looking only at the repo, I was a bit overwhelmed by the possibilities of citelang - in a very positive way :-) As I am not doing research on research software engineering, I was directly looking for an (opinionated), best-practice guide of how to use citelang in a research software package. I understand now that this wasn't the software's intention and you believe that this wouldn't be helpful for users – That's okay for me then and we can still come back to this later during the review process if reviewers ask for it.

> So I'm happy to add what you think is missing, but I don't see that any of the functionality that is provided is missing. Could you tell me what you think is missing? Do you want some sort of a "quick summary" bullet list of things that link to the same examples? Specifically what output are you looking for?

Sorry, i wasn't clear. The example outputs are extensive enough. Then, a very quick summary at the top of [#usage](https://vsoch.github.io/citelang/getting_started/user-guide.html#usage) would be indeed helpful. E.g. what would a research software engineer typically use  **Credit**, **Graph** and **Badges** for and how would one usually present them in a paper or a website -> example links.

> I thought the requirement for JoSS was robust documentation, which I think I have with every example that the software can do, so I apologize if I missed an extra requirement (and I'd like to ask for clarity or a link to what you are looking for).

From what I see, the documentation meets all requirements!

> This is the syntax shown in the links above that is shown with [render](https://vsoch.github.io/citelang/getting_started/user-guide.html#render).

:+1: Maybe a bit of this could be added to the readme.md?
