### 2022-05-20 #4352 @vsoch
> That's okay for me then and we can still come back to this later during the review process if reviewers ask for it.

Yep that works perfectly! Sorry about my wording, we indeed can work on that. I have to be careful because I feel like I can easily offend someone when I talk about this (it already happened in a slack). I think that's what led to "suggests."

> Sorry, i wasn't clear. The example outputs are extensive enough. Then, a very quick summary at the top of [#usage](https://vsoch.github.io/citelang/getting_started/user-guide.html#usage) would be indeed helpful. E.g. what would a research software engineer typically use Credit, Graph and Badges for and how would one usually present them in a paper or a website -> example links.

This sounds perfect! I'll put together a quick "why do I care!" bullet list and add to the README and to the top of the getting started guide.

> Proceeding with the review... @vsoch, I have one or two potential reviewers in mind but I am also interested in your input on potential reviewers, so any suggestions you have are also welcome.

Yes of course! And [here is the spreadsheet](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822) for others - I couldn't easily find it. I can't remember all my old reviews, but I can definitely pick people's names that I don't recognize, here is a random search for "reproducible".

@ gflofst @ daissi @ fabianomenegidio.
