### 2022-05-20 cd9d639
contrib: use specific git log range. Avoid --all. (#36)
* contrib: use specific git log range. Avoid --all.

Because `git log --all` effectively adds REFS in `refs/` to the `git
log` call it will produce output listing commits from all the REFS in
our local repo.

Citelang was getting a list of `--all` commits, and then filtering between
start and end points. Because of this, it could pick up commits that
are not really in the section of history we are interested in - but come
from a fork, or usptream with shared history earlier on.

Call `git log` with a specific commit range so that it direc
[…]
