+++ README.md
@@ -36,6 +36,7 @@ tool to generate a contributors graphic below.
 <table>
   <tr>
     <td align="center"><a href="https://vsoch.github.io"><img src="https://avatars.githubusercontent.com/u/814322?v=4?s=100" width="100px;" alt=""/><br /><sub><b>Vanessasaurus</b></sub></a><br /><a href="https://github.com/vsoch/citelang/commits?author=vsoch" title="Code">💻</a></td>
+    <td align="center"><a href="https://github.com/dtrudg"><img src="https://avatars.githubusercontent.com/u/4522799?v=4?s=100" width="100px;" alt=""/><br /><sub><b>Dave Trudgian</b></sub></a><br /><a href="https://github.com/vsoch/citelang/commits?author=dtrudg" title="Code">💻</a></td>
   </tr>
 </table>
 
