### 2022-05-20 e0b0fb1
adding more documentation and summary of citelang in readme and docs per review
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
