### 2022-05-20 1796989
adding more documentation and summary of citelang in readme and docs … (#38)
* adding more documentation and summary of citelang in readme and docs per review
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
