+++ README.md
@@ -2,28 +2,66 @@
 
 [![PyPI version](https://badge.fury.io/py/citelang.svg)](https://badge.fury.io/py/citelang)
 
-![https://raw.githubusercontent.com/vsoch/citelang/main/docs/assets/img/pypi-citelang.png](https://raw.githubusercontent.com/vsoch/citelang/main/docs/assets/img/pypi-citelang.png)
+Welcome to CiteLang! CiteLang provides methods and graph-based modeling to study software
+ecosystems. You can use CiteLang for your research, or a provided tool to generate
+software graph artifacts, including (but not limited to):
 
-Welcome to CiteLang! This is the first markdown syntax for citing software. Importantly,
-when you use CiteLang to reference software.
-
-1. Generate basic software credit trees (citelang graph, badge, or credit)
+1. Generate basic software credit trees (citelang graph, badge, or markdown credit)
 2. Give credit accounting for dependencies! (see [software-credit.md](software-credit.md))
 3. Actions (automation) for the above!
 
-No - we aren't using DOIs! A manually crafted identifier that a human has to remember to generate,
+For the examples above, we aren't using DOIs! A manually crafted identifier that a human has to remember to generate,
 in addition to a publication or release, is too much work for people to reasonably do. As research
 software engineers we also want to move away from the traditional "be valued like an academic" model.
 We are getting software metadata and a reference to an identifier via a package manager. This means
 that when you publish your software, you should publish it to an appropriate package manager.
 
-⭐️ [Documentation](https://vsoch.github.io/citelang) ⭐️ 
+## Getting Started 
+
+If you want to use CiteLang as an analysis library, jump into the mode detailed ⭐️ [Documentation](https://vsoch.github.io/citelang) ⭐️ 
+or look specifically at the [Python API](https://vsoch.github.io/citelang/getting_started/user-guide.html#python).
+As an example analysis, the [RSEPedia Software Ecosystem](https://rseng.github.io/rsepedia-analysis/) is a completed automated setup that parses and summarizes dependencies across the [Research Software Encyclopedia](https://rseng.github.io/software) weekly, and it's powered by CiteLang! You can do similar analyses or build your own tools using CiteLang. We will provide a small summary of the tools available here.
+
+### Badges
+
+CiteLang [**Badges**](https://vsoch.github.io/citelang/getting_started/user-guide.html#badge) can show an entire credit tree
+for a project:
+
+![https://raw.githubusercontent.com/vsoch/citelang/main/docs/assets/img/pypi-citelang.png](https://raw.githubusercontent.com/vsoch/citelang/main/docs/assets/img/pypi-citelang.png)
+
+or can be generated to be interactive web interfaces as [shown here](https://vsoch.github.io/citelang/_static/example/badge/treemap/index.html).
+
+![https://raw.githubusercontent.com/vsoch/citelang/main/docs/getting_started/img/badge.png](https://raw.githubusercontent.com/vsoch/citelang/main/docs/getting_started/img/badge.png)
+
+See the [badge](https://vsoch.github.io/citelang/_static/example/badge/treemap/index.html) documentation for more examples
+of customizing the look, or level of abstraction. You can automatically generate or update
+a badge for your repository using the provided [GitHub Action](https://vsoch.github.io/citelang/getting_started/user-guide.html#badge-github-action).
+
+### Credit and Graph
+
+If you want to visually show dependency graphs, using [**Credit**](https://vsoch.github.io/citelang/getting_started/user-guide.html#credit)
[…]
