### 2022-05-20 #4352 @vsoch
All done @faroit - I've added more summary / pictures to the README https://github.com/vsoch/citelang#getting-started along with a bullet summary section: https://vsoch.github.io/citelang/getting_started/index.html and added more headers to the GitHub actions to more easily link readers directly to the one they want (it was just one section before).
