### 2022-05-23 #4352 @vsoch
Quick reference to supplement my description above - here is how the Singularity project is using it - https://sylabs.io/2022/05/whats-in-a-singularityce-release/ and  https://twitter.com/Singularity_CE/status/1528785462210469889
