### 2022-05-25 #4352 @vsoch
I haven't worked with any of them - the only connection is that I interviewed the second on RSE Stories a few years ago: https://us-rse.org/rse-stories/2020/robert-haines/.
