### 2022-05-27 #4352 @hainesr
I could _probably_ review this - my reticence is that I don't count Python in the set of ecosystems that I am confident in. I am on leave next week and then very busy in June, but so long as "sometime in June" is an acceptable timeframe, then I can do this.
