### 2022-05-27 #4352 @vsoch
I know him from USRSE (he gives great talks!) but we’ve never worked together, so no I don’t think there is one.
