### 2022-05-29 9fd3808
adding from_file to package_base for single serving usage (#39)
* adding from_file to package_base for single serving usage
also removing deprecated argument from pypi parser and setting default
to not use user site

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
