### 2022-05-31 #4352 @faroit
> @hainesr my reticence is that I don't count Python in the set of ecosystems that I am confident in.

I had a look at the code base and I think a deep knowledge of the python packaging system is likely to be useful. Thanks for your quick reply, though!
