### 2022-06-02 #4352 @rmmilewi
Sure! I'd be happy to do so!

On Wed, Jun 1, 2022, 11:34 PM Fabian-Robert Stöter ***@***.***>
wrote:

> 👋 @rmmilewi <https://github.com/rmmilewi> @gflofst
> <https://github.com/gflofst> @sdruskat <https://github.com/sdruskat>
> @hainesr <https://github.com/hainesr> @yochannah
> <https://github.com/yochannah> - would any of you be willing to review
> this submission for JOSS? The JOSS review process takes place on GitHub and
> focuses on the software and a short paper. We carry out our
> checklist-driven reviews here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> The software under review is CiteLang <https://github.com/vsoch/citelang>.
>
> This issue is a "pre-review" issue in which reviewers are assigned. Once
> sufficient reviewers are recruited we will open a dedicated review issue
> where the review will take place.
>
