### 2022-06-06 #4352 @hainesr
> I had a look at the code base and I think a deep knowledge of the python packaging system is likely to be useful. Thanks for your quick reply, though!

I thought as much! No worries
