### 2022-06-06 vsoch/citelang#41 @faroit
[JOSS] paper bibliography
The way you cite urls

https://github.com/vsoch/citelang/blob/9fd3808f689cd69a6bb03485a39cacb8c1e55a7c/paper/paper.bib#L62

looks correct according to [a number of latex faqs](https://web.archive.org/web/20190428184707/https://texfaq.org/FAQ-citeURL). However, the output isn't generating href links which makes it difficult for readers to visit the cited url.

This may have to fixed on the JOSS template but for now I suggest to use a simpler url citation scheme to enable href linking.

---

issue is part of https://github.com/openjournals/joss-reviews/issues/4352
