### 2022-06-06 #4352 @vsoch
I'm not sure if the bot here is using a cache? I fixed the URLs at the bottom to address https://github.com/vsoch/citelang/issues/41 (see preview attached from Whedon preview service) but it doesn't appear to be reflected above.
[whedon-preview-service.pdf](https://github.com/openjournals/joss-reviews/files/8844387/whedon-preview-service.pdf)
