### 2022-06-06 vsoch/citelang#41 @vsoch
All fixed! Before:

![image](https://user-images.githubusercontent.com/814322/172158877-fd2cc58e-8890-4403-97b6-5b167885a5dc.png)

And after:

![image](https://user-images.githubusercontent.com/814322/172159093-294bd63f-2ed2-4f25-9a55-296d78d85d42.png)

Note that the "editoralbot" preview in the issue linked above seems to be caching the old result - the whedon preview service rendered the above. Thanks for opening the issue @faroit !
