### 2022-06-08 #4458 @editorialbot
[REVIEW]: CiteLang: Modeling the Research Software Ecosystem
**Submitting author:** @vsoch (<a href="http://orcid.org/0000-0002-4387-3819">Vanessa Sochat</a>)
**Repository:** https://github.com/vsoch/citelang
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.33
**Archive:** https://doi.org/10.5281/zenodo.7048920
