### 2022-07-03 #4458 @faroit
@gflofst @rmmilewi please update us about the status of the review.

We aim for reviews to be completed within about 2-4 weeks which is while i will start reminding you using our editorial bot. Let me know if you have problems completing the review or need other help from my side or the authors.
