### 2022-07-08 #4458 @vsoch
We do have automated tests! They are run via GitHub workflows:

 - [test-action](https://github.com/vsoch/citelang/blob/main/.github/workflows/test-action.yml) tests the GitHub actions provided, which also means testing the underlying badge, contrib, and gen commands
 - [main](https://github.com/vsoch/citelang/blob/main/.github/workflows/main.yml) has linting (black and pyflakes) along with runing the test suite at the end - both via pytest and a bash tester that tests the same but via the command line client
 - [build-deploy](https://github.com/vsoch/citelang/blob/main/.github/workflows/build-deploy.yaml) ensures the container build is working that is automatically deployed to the repo
 - [codql](https://github.com/vsoch/citelang/blob/main/.github/workflows/codeql-analysis.yml) is GitHub's code QL analyzer - I'm not convinced yet this is super useful, but I haven't had any errors from it, so maybe that's it.

And here is an example of the "main" test run - if you look under "Run Tests" you'll see lots of output that shows what is going on. https://github.com/vsoch/citelang/runs/6640961948?check_suite_focus=true You can also see the tests here: https://github.com/vsoch/citelang/tree/main/citelang/tests

As an addition, citlelang runs weekly to update the rsepedia, and although it's not an official test, I check that the results look as I'd expect. That's probably a sanity check for a more complex usecase that combines many of the subcommands!

Let me know if you have any questions!
