### 2022-07-08 70eb296
adding test documentation and fixing spack url bug
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
