### 2022-07-08 b9dbb47
Add/test docs (#44)
* adding test documentation and fixing spack url bug

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
