### 2022-07-08 #4458 @vsoch
okay all set! Added a new test section: https://vsoch.github.io/citelang/getting_started/installation.html#testing
