### 2022-09-04 #4458 @faroit
@vsoch Everything looks good from my side. For the next steps, please make a new release of the package and archive it, for example on zenodo. Once you complete those tasks, please report here the version number and doi. Please ensure that the archive's author names and title match exactly those of the paper.pdf.
