+++ citelang/main/package.py
@@ -2,9 +2,9 @@ __author__ = "Vanessa Sochat"
 __copyright__ = "Copyright 2022, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
-import citelang.main.packages as packages
-import citelang.main.endpoints as endpoints
 import citelang.main.cache as cache
+import citelang.main.endpoints as endpoints
+import citelang.main.packages as packages
 from citelang.logger import logger
 
 
