+++ citelang/main/settings.py
@@ -3,16 +3,17 @@ __copyright__ = "Copyright 2022, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
 
-from citelang.logger import logger
-import citelang.defaults as defaults
-import citelang.main.schemas as schemas
-import citelang.utils
+import os
+import re
 import shutil
 
 import jsonschema
-import os
-import re
 import yaml
+
+import citelang.defaults as defaults
+import citelang.main.schemas as schemas
+import citelang.utils
+from citelang.logger import logger
 
 
 class SettingsBase:
