+++ citelang/main/colors.py
@@ -2,8 +2,9 @@ __author__ = "Vanessa Sochat"
 __copyright__ = "Copyright 2022, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
-import colour
 import random
+
+import colour
 
 # Palettes
 
