+++ citelang/main/graph/dot.py
@@ -2,10 +2,10 @@ __author__ = "Vanessa Sochat"
 __copyright__ = "Copyright 2022, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
-from citelang.logger import logger
-
 import random
 import sys
+
+from citelang.logger import logger
 
 from .base import GraphBase
 
