+++ citelang/client/config.py
@@ -2,10 +2,11 @@ __author__ = "Vanessa Sochat"
 __copyright__ = "Copyright 2022, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
-import citelang.main.settings as settings
-import citelang.defaults as defaults
-from citelang.logger import logger
 import sys
+
+import citelang.defaults as defaults
+import citelang.main.settings as settings
+from citelang.logger import logger
 
 
 def main(args, parser, extra, subparser):
