+++ citelang/utils/__init__.py
@@ -1,15 +1,6 @@
-from .string import update_nested, get_terminal_pad
-from .terminal import (
-    clone,
-    run_command,
-    get_installdir,
-    get_time_now,
-    confirm_action,
-    confirm_uninstall,
-)
 from .fileio import (
-    get_tmpfile,
     get_tmpdir,
+    get_tmpfile,
     mkdir_p,
     print_json,
     read_file,
@@ -18,4 +9,13 @@ from .fileio import (
     workdir,
     write_file,
     write_json,
+)
+from .string import get_terminal_pad, update_nested
+from .terminal import (
+    clone,
+    confirm_action,
+    confirm_uninstall,
+    get_installdir,
+    get_time_now,
+    run_command,
 )
