### 2022-09-04 1ee9a78
preparing for 0.0.33 release for JoSS paper (#46)
* preparing for 0.0.33 release for JoSS paper
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
