+++ citelang/main/endpoints.py
@@ -2,12 +2,13 @@ __author__ = "Vanessa Sochat"
 __copyright__ = "Copyright 2022, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
-from citelang.logger import logger
+import sys
+from datetime import datetime
+
 import citelang.defaults as defaults
 import citelang.main.http as http
 import citelang.main.result as results
-from datetime import datetime
-import sys
+from citelang.logger import logger
 
 # Registered endpoints (populated on init)
 registry = {}
