+++ citelang/main/graph/cypher.py
@@ -3,9 +3,9 @@ __copyright__ = "Copyright 2022, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
 
-from citelang.logger import logger
-
 import sys
+
+from citelang.logger import logger
 
 from .base import GraphBase
 
