+++ citelang/main/packages/cran_description.py
@@ -5,6 +5,7 @@ __license__ = "MPL 2.0"
 import re
 
 from citelang.logger import logger
+
 from .base import PackagesFromFile
 
 
