+++ citelang/main/packages/npm_packages.py
@@ -3,7 +3,6 @@ __copyright__ = "Copyright 2022, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
 import json
-
 import re
 
 from .base import PackagesFromFile
