+++ citelang/main/packages/python_base.py
@@ -2,8 +2,9 @@ __author__ = "Vanessa Sochat"
 __copyright__ = "Copyright 2022, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
-from citelang.logger import logger
 import re
+
+from citelang.logger import logger
 
 from .base import PackagesFromFile
 
