### 2022-09-04 10c470b
adding citation.cff and doi badge
Signed-off-by: vsoch <vsoch@users.noreply.github.com>
