+++ CITATION.cff
@@ -1,0 +1,10 @@
+cff-version: 1.1.0
+message: "If you use this software, please cite it as below."
+authors:
+  - family-names: Sochat
+    given-names: Vanessa
+    orcid: https://orcid.org/0000-0002-4387-3819
+title: "CiteLang: Modeling the Research Software Ecosystem"
+version: 0.0.33
+date-released: 2022-09-04
+            
