### 2022-09-04 #4458 @vsoch
okay DOI is created, added to the repository along with a CITATION.cff! https://github.com/vsoch/citelang

I think Zenodo might be lagged in actually creating the page for it - there is a notification at the top about uploading issues. So I can't edit the metadata and ensure the author/title are OK, but I'll keep checking it today and update when I can. I can also ping you here when it is fully formed if you like!
