### 2022-09-05 #4458 @arfon
@vsoch  – your paper is now accepted and published in JOSS :zap::rocket::boom: This is such a neat piece of work – I'm hoping I get to use it!
