### 2022-09-05 #4458 @faroit
@vsoch congrats and thanks for all your patience! I am not very happy with duration of the review process so I want to apologize for the delay also were coming from my side. I hope you will choose JOSS again and I will try harder to speed up the process next time.
